refactor(backend): extract sync watch service
